fk search cell
