+++ apps/forms/scaffolds/grid-editor/components/query/xsb/index.ts
@@ -1,2 +1,3 @@
 export * from "./xsb-sql-fk-search-input";
 export * from "./xsb-sql-literal-input";
+export * from "./xsb-search-table-sheet";
