+++ apps/forms/scaffolds/grid/cells/index.ts
@@ -3,3 +3,4 @@ export * from "./column-header-cell";
 export * from "./file-cell";
 export * from "./json-cell";
 export * from "./richtext-cell";
+export * from "./xsb-fk-cell";
